Allow configuration of CERN Virtual Machine
Aldd module to manage the configuration files for the CERN Virtual
Machine.  The configuration files are placed in /etc/cernvm.  Should
olnly be applied on CernVMs. Configurable options are the version of
cvmfs to use, whether CernVM should acts as batch node, desktop,
etc., what experiment the cernvm should be set up for, which cvmfs
repositories will be made avaiable, and where to place the
configuration files.

Changes:
   * new file:   cernvm/manifests/init.pp
   * new file:   cernvm/templates/site.conf.erb
   * modified:   gce_node/manifests/init.pp